# api/openapi.yaml
openapi: 3.1.0
info:
  title: Task Manager API
  version: 1.0.0

servers:
  - url: '/'

tags:
  - name: Tasks
  - name: TaskWatchers

paths:
  /api/v1/task-watcher:
    post:
      tags:
        - TaskWatchers
      summary: Add a watcher to a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskWatcher'
      responses:
        '201':
          description: Task watcher added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskWatcher'
        '404':
          description: Task not found
        '409':
          description: Watcher already exists for this task

  /api/v1/tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks from Api first yaml
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      tags:
        - Tasks
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        dueDate:
          type: string
          format: date
      required:
        - title
        - status

    TaskStatus:
      type: string
      enum:
        - OPEN
        - IN_PROGRESS
        - DONE

    TaskWatcher:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        taskId:
          type: string
          format: uuid
        emailId:
          type: string
          format: email
      required:
        - taskId
        - emailId